Remove unused font families.
Put all the declarations in a separate file, so I can find them again later.
Also fix the medium weight, which should be 500, not normal (400), and remove
some redundant font weight specs in screen.css. While at it, switch to the
external link font awesome icon instead of an arrow. The lite version would
look better, but for now use this free version.