Refine aws sync in publish script.
Need to be more precise in establishing media type mappings. Also, turns out
--delete only affects the files matched, not those unmatched, so there had
been a lot of detritus left over. Now we try to cover all the appropriate
bases, and exclude everything we included in the final call to let awscli make
guesses (which it does fine for JSON, PNG, JPEG, etc.).