Footnote updates
Replace the space between a link and a footnote with a zero-width space. Both
work around yuin/goldmark#143, but the zero-width space is more aesthetically
pleasing.

Also remove a content rewrite of the footnote return link to move it into a
paragraph; Goldmark does not seem to have this issue.