Remove "tap to zoom"
For now. Because only images above the body copy (that is, defined by the
`image:` config in the front matter) are zoomable. This is because those in
the copy are inside `<div class="text">`, which has a max size for the text
that we don't want to exceed. Consider reverting this change if I figure out a
way to have zoomable images inside the body copy, presumably by moving them
outside the text div, or else by letting the text be full width and set the
max width for its children. Another time, though.